missed power_limit

+++ include/amd_smi/amdsmi.h
@@ -1096,7 +1096,7 @@ typedef struct {
     uint64_t gfx_voltage;           //!< GFX voltage measurement in mV {@linux_bm} or V {@host}
     uint64_t soc_voltage;           //!< SOC voltage measurement in mV {@linux_bm} or V {@host}
     uint64_t mem_voltage;           //!< MEM voltage measurement in mV {@linux_bm} or V {@host}
-    uint64_t power_limit;           //!< The power limit in W {@linux_bm}, Linux only
+    uint32_t power_limit;           //!< The power limit in W {@linux_bm}, Linux only
     uint64_t reserved[18];
 } amdsmi_power_info_t;
 /**
